requirements.txt
asgiref==3.7.2
dj-database-url==0.2.2
Django==2.2.24
django-adminlte3==0.1.6
django-cors-headers==3.0.2
django-rest-passwordreset==1.1.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
gunicorn==22.0.0
numpy==1.26.1
packaging==24.0
pandas==2.1.1
psycopg2==2.9.9
PyJWT==1.7.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.1.0
pytz==2024.1
sendgrid==6.0.5
six==1.12.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
whitenoise==6.6.0
numpy==1.26.1
